split up 'crates/search/src/generation/text_search'
