+++ crates/search/src/generation/text_search/mod.rs
@@ -10,31 +10,20 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-use std::{any::Any, cmp::min, collections::HashMap, fmt::Formatter, sync::Arc};
+use std::{cmp::min, collections::HashMap, sync::Arc};
 
 use crate::{SEARCH_SCORE_COLUMN_NAME, SEARCH_VALUE_COLUMN_NAME};
 use arrow::{
     array::RecordBatch,
-    datatypes::{Field, Schema, SchemaRef},
+    datatypes::{Field, Schema},
     error::ArrowError,
 };
 use arrow_json::reader::Decoder;
 use async_stream::stream;
 use async_trait::async_trait;
 use datafusion::{
-    catalog::{Session, TableProvider},
-    common::Constraints,
-    datasource::TableType,
-    error::{DataFusionError, Result as DataFusionResult},
-    execution::SendableRecordBatchStream,
-    logical_expr::{TableProviderFilterPushDown, sqlparser::ast::Expr},
-    physical_expr::EquivalenceProperties,
-    physical_plan::{
-        DisplayAs, DisplayFormatType, ExecutionPlan, Partitioning, PlanProperties,
-        execution_plan::{Boundedness, EmissionType},
-        stream::RecordBatchStreamAdapter,
-    },
-    prelude::Expr as LogicalExpr,
+    error::DataFusionError, execution::SendableRecordBatchStream,
+    logical_expr::sqlparser::ast::Expr, physical_plan::stream::RecordBatchStreamAdapter,
     sql::sqlparser::ast::Ident,
 };
 
@@ -56,6 +45,9 @@ use super::{
 };
 
 static DEFAULT_BATCH_SIZE: usize = 100;
+
+pub mod exec;
+pub mod table;
 
 #[derive(Debug, Snafu)]
 pub enum Error {
@@ -439,209 +431,6 @@ fn tantivy_json_to_arrow_decoder(hits: &[Value]) -> std::result::Result<Decoder,
     decoder.serialize(hits)?;
 
     Ok(decoder)
-}
-
-/// An implementation of [`TableProvider`] based on a known query on a [`FullTextSearch`] index.
-#[derive(Clone, Debug)]
-pub struct FullTextSearchTable {
-    index: FullTextSearch,
-    query: String,
-    default_limit: usize,
-}
-
-impl FullTextSearchTable {
-    #[must_use]
-    pub fn new(index: FullTextSearch, query: String) -> Self {
-        Self {
-            index,
-            query,
-            default_limit: DEFAULT_BATCH_SIZE,
-        }
-    }
-
-    #[must_use]
-    pub fn with_limit(mut self, limit: usize) -> Self {
-        self.default_limit = limit;
-        self
-    }
-}
-#[async_trait]
-impl TableProvider for FullTextSearchTable {
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-
-    fn schema(&self) -> SchemaRef {
-        let tantivy_schema = self.index.idx.schema();
-
-        let fields = self
-            .index
-            .all_columns()
-            .iter()
-            .filter_map(|field_name| {
-                let f = tantivy_schema.get_field(field_name).ok()?;
-                let entry = tantivy_schema.get_field_entry(f);
-                let data_type = tantivy_to_arrow_type(entry.field_type())?;
-                Some(Field::new(field_name, data_type, false))
-            })
-            .chain([Field::new(
-                SEARCH_SCORE_COLUMN_NAME,
-                arrow::datatypes::DataType::Float64,
-                false,
-            )])
-            .collect::<Vec<_>>();
-
-        Arc::new(Schema::new(fields))
-    }
-
-    fn constraints(&self) -> Option<&Constraints> {
-        // TODO primary keys
-        None
-    }
-
-    fn table_type(&self) -> TableType {
-        TableType::Base
-    }
-
-    async fn scan(
-        &self,
-        _state: &dyn Session,
-        projection: Option<&Vec<usize>>,
-        filters: &[LogicalExpr],
-        limit: Option<usize>,
-    ) -> DataFusionResult<Arc<dyn ExecutionPlan>> {
-        Ok(Arc::new(
-            FullTextSearchTableExec::try_new(
-                self.clone(),
-                projection,
-                filters.to_vec(),
-                limit.unwrap_or(self.default_limit),
-            )
-            .map_err(DataFusionError::from)?,
-        ))
-    }
-
-    fn supports_filters_pushdown(
-        &self,
-        _filters: &[&LogicalExpr],
-    ) -> DataFusionResult<Vec<TableProviderFilterPushDown>> {
-        Ok(vec![])
-    }
-}
-
-pub struct FullTextSearchTableExec {
-    tbl: FullTextSearchTable,
-    filters: Vec<LogicalExpr>,
-    limit: usize,
-    plan_properties: PlanProperties,
-}
-
-impl FullTextSearchTableExec {
-    pub fn try_new(
-        tbl: FullTextSearchTable,
-        projection: Option<&Vec<usize>>,
-        filters: Vec<LogicalExpr>,
-        limit: usize,
-    ) -> Result<Self, ArrowError> {
-        let schema = match projection {
-            Some(proj) => Arc::new(tbl.schema().project(proj.as_slice())?),
-            None => Arc::clone(&tbl.schema()),
-        };
-
-        Ok(Self {
-            tbl,
-            filters,
-            limit,
-            plan_properties: PlanProperties::new(
-                EquivalenceProperties::new(schema),
-                Partitioning::UnknownPartitioning(1),
-                EmissionType::Incremental,
-                Boundedness::Bounded,
-            ),
-        })
-    }
-}
-
-impl std::fmt::Debug for FullTextSearchTableExec {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("FullTextSearchTableExec")
-            .field("tbl", &self.tbl)
-            .field("filters", &self.filters)
-            .field("limit", &self.limit)
-            .finish_non_exhaustive()
-    }
-}
-impl DisplayAs for FullTextSearchTableExec {
-    fn fmt_as(&self, _t: DisplayFormatType, f: &mut Formatter) -> std::fmt::Result {
-        write!(f, "FullTextSearchTableExec q={}", self.tbl.query)
-    }
-}
-
-impl ExecutionPlan for FullTextSearchTableExec {
-    fn name(&self) -> &'static str {
-        "FullTextSearchTableExec"
-    }
-
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-
-    fn properties(&self) -> &PlanProperties {
-        &self.plan_properties
-    }
-
-    fn children(&self) -> Vec<&Arc<dyn ExecutionPlan>> {
-        vec![]
-    }
-
-    fn with_new_children(
-        self: Arc<Self>,
-        _children: Vec<Arc<dyn ExecutionPlan>>,
-    ) -> DataFusionResult<Arc<dyn ExecutionPlan>> {
-        Ok(Arc::clone(&self) as Arc<dyn ExecutionPlan>)
-    }
-
-    fn execute(
-        &self,
-        _partition: usize,
-        _context: Arc<datafusion::execution::TaskContext>,
-    ) -> DataFusionResult<SendableRecordBatchStream> {
-        let idx = self.tbl.index.clone();
-        let schema = self.schema();
-        let limit = self.limit;
-        let query = self.tbl.query.clone();
-        let s = stream! {
-        // TODO: Support filters.
-            match idx
-                .search(query, &[], &[], limit)
-                .await
-                .map_err(|e| DataFusionError::Plan(format!("Failed to prepare full text search: {e}"))) {
-                Ok(mut stream) => {
-                    while let Some(item) = stream.next().await {
-                        match item {
-                            Err(e) => yield Err(e),
-                            Ok(rb) => {
-                                // Apply projection, as per `self.schema()`, to record batch from FTS.
-                                let proj = rb.schema().fields().iter().enumerate().filter_map(|(i, f)| {
-                                    if schema.column_with_name(f.name()).is_some() {
-                                        Some(i)
-                                    } else {
-                                        None
-                                    }
-                                }).collect::<Vec<_>>();
-                                yield rb.project(proj.as_slice()).map_err(DataFusionError::from)
-                            }
-                        }
-                    }
-                },
-                Err(e) => {
-                    yield Err(e);
-                    return;
-                }
-            }
-        };
-        Ok(Box::pin(RecordBatchStreamAdapter::new(self.schema(), s)))
-    }
 }
 
 fn tantivy_to_arrow_type(t: &FieldType) -> Option<arrow::datatypes::DataType> {
